requirements.txt
APScheduler==3.10.1
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dataclasses-json==0.5.8
Django==4.2.9
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.7
marshmallow==3.26.1
marshmallow-enum==1.5.1
mypy_extensions==1.1.0
packaging==25.0
phonepe-sdk==2.1.4
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
typing-inspect==0.9.0
typing_extensions==4.13.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==1.26.12
